- add FreeBSD targets to all frameworks - brought X11 back to current for Darwin and FreeBSD - removed CairoCacheImage, added O2Surface_cairo - separated TTFFont into O2FontState_cairo/O2Font_FT/TTFFont - renamed CairoContext O2Context_cairo - added O2Context_builtin_FT which implements Onyx2D+FreeType - moved font queries to X11Display - added NSPlatform_bsd.m to AppKit - encoding change/fix(?) in NSColor.m - moved showText: into abstract class O2Context from _gdi classes - fix for ObjCModule.m to get executable path

+++ AppKit/NSColorPicker.subproj/NSColorPickerSliders.m
@@ -12,7 +12,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 
 @implementation NSColorPickerSliders
 
--initWithPickerMask:(int)mask colorPanel:(NSColorPanel *)colorPanel {
+-initWithPickerMask:(NSUInteger)mask colorPanel:(NSColorPanel *)colorPanel {
     [super initWithPickerMask:mask colorPanel:colorPanel];
 
     NSSize size=[_grayscaleConstantsMatrix cellSize];
